# configuration/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Techincal Analysis Indicators",
    "description": "",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://taindicators-2ryqr.ondigitalocean.app/"
    }
  ],
  "tags": [
    {
      "name": "health_check",
      "description": "check if the app is working"
    },
    {
      "name": "indicators"
    }
  ],
  "paths": {
    "/health_check": {
      "get": {
        "tags": [
          "health_check"
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/simple_moving_average/{coin}/{time}": {
      "get": {
        "tags": [
          "indicators"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "coin",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Symbol of the coin (e.g. BTC)"
          },
          {
            "in": "path",
            "name": "time",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "time period; accepted values = [`1h` `3h` `12h` `24h` `7d` `30d` `3m` `1y` `3y` `5y`]"
          }
        ],
        "description": "A simple moving average (SMA) is an arithmetic moving average calculated by adding recent prices and then dividing that figure by the number of time periods in the calculation average.",
        "responses": {
          "200": {
            "description": "response example",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "format": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "number",
                      "format": "number",
                      "example": 37386.791344726465
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/fibonacci_retracement/{coin}/{trend}": {
      "get": {
        "tags": [
          "indicators"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "coin",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Symbol of the coin (e.g. BTC)"
          },
          {
            "in": "path",
            "name": "trend",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "accepted values for `market` = [ `U` or `UPTREND` | `D` or `DOWNTREND` ]"
          }
        ],
        "description": "Fibonacci retracement levels—stemming from the Fibonacci sequence—are horizontal lines that indicate where support and resistance are likely to occur. Each level is associated with a percentage. The percentage is how much of a prior move the price has retraced.",
        "responses": {
          "200": {
            "description": "the response example is not accurate",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "format": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "percentage": {
                            "type": "string",
                            "format": "string",
                            "example": "38.2%"
                          },
                          "value": {
                            "type": "number",
                            "format": "number",
                            "example": 20353.453291087826
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/fibonacci_extension/{coin}/{trend}": {
      "get": {
        "tags": [
          "indicators"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "coin",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Symbol of the coin (e.g. BTC)"
          },
          {
            "in": "path",
            "name": "trend",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "accepted values for `market` = [ `U` or `UPTREND` | `D` or `DOWNTREND` ]"
          }
        ],
        "description": "Fibonacci extensions are a tool that traders can use to establish profit targets or estimate how far a price may travel after a pullback is finished",
        "responses": {
          "200": {
            "description": "the response example is not accurate",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "format": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "percentage": {
                            "type": "string",
                            "format": "string",
                            "example": "38.2%"
                          },
                          "value": {
                            "type": "number",
                            "format": "number",
                            "example": 20353.453291087826
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/rsi/{coin}": {
      "get": {
        "tags": [
          "indicators"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "coin",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Symbol of the coin (e.g. BTC)"
          }
        ],
        "summary": "14 days period",
        "description": "The Relative Strength Index (RSI), developed by J. Welles Wilder, is a momentum oscillator that measures the speed and change of price movements. The RSI oscillates between zero and 100. Traditionally the RSI is considered overbought when above 70 and oversold when below 30.",
        "responses": {
          "200": {
            "description": "example",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "format": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "number",
                      "format": "number",
                      "example": 60.34680279684091
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/aroon_oscillator/{coin}": {
      "get": {
        "tags": [
          "indicators"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "coin",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Symbol of the coin (e.g. BTC)"
          }
        ],
        "summary": "25 days period",
        "description": "The Aroon Oscillator is a trend-following indicator that uses aspects of the Aroon Indicator (Aroon Up and Aroon Down) to gauge the strength of a current trend and the likelihood that it will continue.",
        "responses": {
          "200": {
            "description": "example",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "format": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "number",
                      "format": "number",
                      "example": -32
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/stochastic_oscillator/{coin}": {
      "get": {
        "tags": [
          "indicators"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "coin",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Symbol of the coin (e.g. BTC)"
          }
        ],
        "summary": "14 days period",
        "description": "A stochastic oscillator is a momentum indicator comparing a particular closing price of a security to a range of its prices over a certain period of time. The sensitivity of the oscillator to market movements is reducible by adjusting that time period or by taking a moving average of the result.",
        "responses": {
          "200": {
            "description": "example",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "format": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "number",
                      "format": "number",
                      "example": 89.3414921071439
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/exponential_moving_average/{coin}": {
      "get": {
        "tags": [
          "indicators"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "coin",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Symbol of the coin (e.g. BTC)"
          }
        ],
        "summary": "20 days period",
        "description": "An exponential moving average (EMA) is a type of moving average (MA) that places a greater weight and significance on the most recent data points",
        "responses": {
          "200": {
            "description": "example",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "format": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "number",
                      "format": "number",
                      "example": 16771.16256415485
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}